Cleanup and unify desktop and web samples
